requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
astunparse==1.6.3
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.14
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
gotrue==2.12.3
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
keras==3.11.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.6.3
namex==0.1.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-contrib-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==10.3.0
postgrest==1.0.2
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
realtime==2.4.3
requests==2.32.4
rich==14.1.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
